requirements.txt
huggingface_hub>=0.20.0
numpy>=1.24.0
onnx>=1.14.0
onnxruntime-gpu>=1.17.0
pillow>=10.0.0
tqdm>=4.66.0
